Fix minc handling in GlobalIndexUidAggregator
If a number of mutations that add protocol buffers with only removed
UIDs are minor compacted, then those keys will be dropped since the
count is <= 0 (given that the protocol buffer has only removed UIDs)
and therefore the propogateKey method will return false causing the
PropogatingIterator to drop the value. This what is desired for a full
major compaction since we know the protocl buffer has combined
everything known for a given key and doesn't have any UIDs, only
removals. However, for a minor compaction, this is very wrong. If the
sequence of events is as follows, then removals are lost:

1. Insert a document
2. Flush/compact the table
3. Insert the removal of the document (-1 counts in the value protocol
   buffers). If this is done with live ingest, then these keys are in
   memory.
4. Flush the table to cause a minor compaction of the removals. The
   keys will all be dropped.

This change fixes that bug by ensuring that a key is always propagated
when the propogate flag is set to true.